Files renamed and spliter separated in another directory
